% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{interpret_capital}
\alias{interpret_capital}
\title{Interpret Capital Requirements}
\usage{
interpret_capital(startup_costs, personal_runway)
}
\arguments{
\item{startup_costs}{A list as returned by \code{\link[=calc_startup_costs]{calc_startup_costs()}}.}

\item{personal_runway}{A list as returned by \code{\link[=calc_personal_runway]{calc_personal_runway()}}.}
}
\value{
A character string of narrative text.
}
\description{
Generates narrative text describing startup capital and personal runway
requirements, framed around what financing decision they imply.
}
